+++ src/org/ods/quickstarter/Context.groovy
@@ -1,13 +1,29 @@
 package org.ods.quickstarter
 
-import com.cloudbees.groovy.cps.NonCPS
+import org.ods.util.Logger
+import org.ods.util.IPipelineSteps
+import org.ods.util.PipelineSteps
+
+import org.ods.services.OpenShiftService
+import org.ods.services.ServiceRegistry
 
 class Context implements IContext {
 
-    private final Map config
+    private final Logger logger
+    private IPipelineSteps steps
+    private final def script
+    private OpenShiftService openShiftService
 
-    Context(Map config) {
+    Context(Map config, Logger logger, def script) {
         this.config = config
+        this.logger = logger
+        this.steps = new PipelineSteps(script)
+    }
+
+    def assemble() {
+        if (!config.containsKey('appDomain')) {
+            getOpenshiftApplicationDomain()
+        }
     }
 
     @NonCPS
@@ -145,4 +161,26 @@ class Context implements IContext {
         config.odsGitRef
     }
 
+    @NonCPS
+    String getAppDomain() {
+        config.appDomain
+    }
+
+    void setAppDomain() {
+        getOpenshiftApplicationDomain()
+    }
+
+    void getOpenshiftApplicationDomain () {
+        if (!config.projectId) {
+            logger.debug('Could not get application domain, as no projectId is available')
+            return ''
+        }
+        logger.startClocked("${config.componentId}-get-oc-app-domain")
+
+        this.openShiftService = new OpenShiftService(steps, logger)
+        config.appDomain = openShiftService.getApplicationDomain("${config.projectId}-cd")
+
+        logger.debugClocked("${config.componentId}-get-oc-app-domain")            
+
+    }
 }