+++ src/org/ods/quickstarter/IContext.groovy
@@ -71,4 +71,10 @@ interface IContext {
     // alias for odsGitRef
     String getGitBranch()
 
+    // get app domain of the Openshift cluster
+    String getAppDomain()
+
+    // set app domain of the Openshift cluster
+    void setAppDomain()
+
 }