+++ src/org/ods/quickstarter/Pipeline.groovy
@@ -148,7 +148,8 @@ class Pipeline implements Serializable {
             annotations: config.annotations,
         ) {
             script.node(podLabel) {
-                IContext context = new Context(config)
+                IContext context = new Context(config, logger, script)
+                context.assemble()
                 block(context)
             }
         }